requirement.txt
asgiref==3.5.0
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
zipp==3.8.0
